update mysqlclient (#4)
The reconnect functionality was changd in the MySQL/MariaDB client
libraries. This caused the disappearance of the reconnect field in the
header file.
Adding that field back doesn't bring back the functionality. Instead
this might cause driver problems in the application.

Therefore updating the MySQL python driver is the better option.