fix error handling for invalid cookies
It happens every now and then, that cookies are being denied by the
SecureCookieSession, but we do not get a reason.
In these cases, an abort was sent, which resulted in unhandled output
and then stack traces and crashes in python handlers.

Now these error cases get reported in the same json style and should
lead to less problems on the client side.
It was tested with the available ndcli and instead of crashing it will
report the error and return a proper exit code.