add release workflow
This adds a release workflow compatible with github actions. This way we
can build releases whenever we need to.
This will produce single source files and currently CentOS8 packages.

For CentOS8 the dim rpm will not install into the standard paths,
because CentOS8 has too old python packages available at the moment.

Packages for Debian need to be built yet.