add dim-web workflow
This adds the dim-web workflow to package dim-web. In addition it also
adds a small fix to the dim workflow to avoid both tags working for
dim.

Some small adjustment needed to be made, like the renaming of the python
package. Debian didn't like it so much, that the python package name was
cas when it should have been dim-web.