provide automated packaging workflows (#26)
* add release workflow

This adds a release workflow compatible with github actions. This way we
can build releases whenever we need to.
This will produce single source files and currently CentOS8 packages.

For CentOS8 the dim rpm will not install into the standard paths,
because CentOS8 has too old python packages available at the moment.

Packages for Debian need to be built yet.

* add basic debian package building

With the added functionality the debian buster packages are now built
too for python3 only.

* add github release creation

This will create a github release and attach all files to the release.
It will not be made public to avoid problems.

* dimclient - split of workflow for dimclient release

This splits of the workflow for the dimclient release process. It
catches tags for dimclient-* and based on that builds the release.

* ndcli - split ndcli release workflow

This splits out the ndcli release workflow from the main workflow file.
It catches tags like ndcli-* and starts the release process.

* dim - split of dim release workflow

This splits of the dim release workflow from the main build file. It
catches dim-* tags and starts the release process.

* remove global release build

This workflow was the basis for everything else, but is not needed
anymore, so we can delete that.
It will live on in the dim, ndcli and dimclient workflows.

* dim - add centos7 release build