Package dim web (#28)
* add release workflow

This adds a release workflow compatible with github actions. This way we
can build releases whenever we need to.
This will produce single source files and currently CentOS8 packages.

For CentOS8 the dim rpm will not install into the standard paths,
because CentOS8 has too old python packages available at the moment.

Packages for Debian need to be built yet.

* add basic debian package building

With the added functionality the debian buster packages are now built
too for python3 only.

* add github release creation

This will create a github release and attach all files to the release.
It will not be made public to avoid problems.

* add dim-web workflow

This adds the dim-web workflow to package dim-web. In addition it also
adds a small fix to the dim workflow to avoid both tags working for
dim.

Some small adjustment needed to be made, like the renaming of the python
package. Debian didn't like it so much, that the python package name was
cas when it should have been dim-web.