dim, ndcli - add attribute output to list pools
This makes it possible to list custom attributes when listing pools.

The `list pools` command got a new flag `-a` or `--attributes`, the same
as with list ips, which takes a list of attributes to return.

The API endpoint ippools_list got the new parameter attributes, which
takes the list of fields to return.