dim - fix compatibility with older ndcli clients
When introducing the -a option to select pool attributes in
24625243b077e57013885b737dd64fae458f8d8b a new parameter attributes was
added to ippool_list with the default being None.

This causes a problem with older clients. Attributes is tried to be
iterated over when it is checked against the reserved keywords.
But older clients do not send the attribute parameter, which causes it
to be None and therefore causing

	ERROR - 'NoneType' object is not iterable

When we switch the default to being an empty list, the output will be as
before and even older clients can get the same output.