ndcli - remove comment import for zones
ndcli is able to import comments in zone files, which look like this:

	foo A 127.0.0.1 ; a comment

WIth dnspython 2.1 an internal API changed and now the patch to
dnspython via ndcli doesn't work anymore.

This means, that comment import from bind files for each record does not
work anymore, but fixing that problem would take much more time to
figure out correctly.
And if we do not figure out how to work around the issue, it will throw
us back with compatibility even more.