dim-testsuite - get testsuite running
There were a lot of errors because of various forms of str and byte
incompatibilties. I wasn't able to track them all down, but now it is
working and it will need a lot of work in the future.

At the same time the ndcli call was using the library, which was
replaced with subprocess calls, because I wasn't able to get proper
results from the library calls.