dim - python2 to 3 conversion changes
These are all conversion changes made by the 2to3 tool.

In rpc.py#1462 is also a small change in the sqlalchemy syntax, because
as it was, that query did not work. The profile attribute in the
filter_by clause was tried against the name string.
Converting that to a filter clause made the query work.