dim-testsuite - fix byte/str problems
There were a lot of problems when comparing the result with the expected
result, because bytes and str types were all over the place. Sometimes
even int was involved.
This tries to fix most of the places, but I am sure, this is still not
the last place. Wouldn't it be nice if a compiler could tell me at once
what is wrong instead of me figuring out every line by myself?