dim - fix history for subnets added to pools
Somehow in all the SQLAlchemy changes the event system changed in a way,
that adding a subnet to a pool didn't trigger the history entry anymore.

With these commits the history entry is created as before. If there is a
better way, then please fix it. I couldn't find a better way.