make dim testsuite work (#50)
Group of changes to make DIM work with python 3. The whole system will probably be numbered DIM 4.0

* ndcli - remove comment import for zones

ndcli is able to import comments in zone files, which look like this:

	foo A 127.0.0.1 ; a comment

With dnspython 2.1 an internal API changed and now the patch to
dnspython via ndcli doesn't work anymore.

This means, that comment import from bind files for each record does not
work anymore, but fixing that problem would take much more time to
figure out correctly.
And if we do not figure out how to work around the issue, it will throw
us back with compatibility even more.

* move tests into correct directory

These are needed by the runtest.py script as it contains a couple
helpers.

* dim-testsuite - add Makefile to prepare environment

* dim-testsuite - get testsuite running

There were a lot of errors because of various forms of str and byte
incompatibilties. I wasn't able to track them all down, but now it is
working and it will need a lot of work in the future.

At the same time the ndcli call was using the library, which was
replaced with subprocess calls, because I wasn't able to get proper
results from the library calls.

* dim - replace pycrypto

According to https://github.com/pycrypto/pycrypto/issues/173, pycrypto
is not maintained since 2015. So replace it with the fork to stop a
couple error messages that popped up during tests.

* dim-testsuite - start the webserver when running the tests

This makes starting the daemon beforehand unnecessary and makes testing
a tad bit faster.

* dim-testsuite - add README and Makefile

This should help to get the tests up and running.

* dim-testsuite - fix stdin errors

When the _ndcli method was reimplemented stdin was forgotten to be
added. Now that it is back in, we can get results from the zone import
scripts.

* dim - add missing order_by clause

The order by is important to keep the order of the zone list results
ordered by name.

* dim - mark crypto string as bytes array

* dim - replace missing keydata field

The keydata field went missing when upgrading the pycrypto dependency to
pycryptodome. This was the original content.

* dim - python2 to 3 conversion changes

These are all conversion changes made by the 2to3 tool.

In rpc.py#1462 is also a small change in the sqlalchemy syntax, because
as it was, that query did not work. The profile attribute in the
filter_by clause was tried against the name string.
Converting that to a filter clause made the query work.

* dim, ndcli - fix argument garbage

This didn't work since forever.

The parameter list didn't require the most important arguments, which
caused the first named argument to be a dict instead of the expected
type.
With this change, we require the parameter to exist at least and then
also add it to ndcli.

* dim - enforce argument type

The argument for islice was ecpected to be an integer but entered as an
integer.
This caused then errors further down the line.

* dim - avoid floats

When creating the range array, the range was split. This computation
resulted in a float, which then caused errors further down the line,
when an IP should have been generated.

To avoid that problem, use floor to make sure we always get an integer
out of the computation.

* dim - fix python3 type handling

In python3 a list of characters is a list of integers. In python2 it was
a list of strings.

So we don't need the ord call anymore.

* dim - fix string and byte incompatibility

In python3 it isn't possible anymore to concat a string to bytes. So
make everything into bytes and make it work that way.

* dim-testsuite - make README nicer

* dim-testsuite - make clear how to run the server

runtest.py can start the server itself, but will not try killing what
already exists.

* dim-testsuite - fix byte/str problems

There were a lot of problems when comparing the result with the expected
result, because bytes and str types were all over the place. Sometimes
even int was involved.
This tries to fix most of the places, but I am sure, this is still not
the last place. Wouldn't it be nice if a compiler could tell me at once
what is wrong instead of me figuring out every line by myself?

* dim-testsuite: resolve match error

When a parenthesis was present in a string, that was used in a regex, it
was causing errors when using the regex.
With the escape in place, this causes less problems.

Also fix some more string/byte errors.

* dim-testsuite - fix order of output

The order of the output was wrong, which caused the tests to not
succeed.

* dim-testsuite - make more places binary

There were still some places that were strings, instead of binary, which
caused the regex library to crash.

* dim-testsuite - simplify comparison of lines

The result was matched over and over with what was expected. Instead of
wasting time with loops, the output should be constant to what is
expected.

So in this commit the lopps were simplified, which also reduces the
runtime of the testsuite a bit.

* dim - add order by options

Until now, the order of returned rows was sometimes random, which
confused the tests.
WIth this, the order is predefined and should be much more stable.

* ndcli - fix order of returned rows

Somehow, the order of returned users is sometimes random, even though it
should be sorted in the middleware.
So instead enforce the order in the client to make the checks go green.

* dim-testsuite - fix the last string/byte problems

This should be now the last couple issues fixed.

* dim-testsuite - remove trailing whitespace

* dim-testsuite - fix test ttl vs serial

The ttl is always set before the serial, so the history must always be
this way.

* dim - fix history for subnets added to pools

Somehow in all the SQLAlchemy changes the event system changed in a way,
that adding a subnet to a pool didn't trigger the history entry anymore.

With these commits the history entry is created as before. If there is a
better way, then please fix it. I couldn't find a better way.

* dim-testsuite - print what is wrong

Sometimes, there are such small issues wrong, that it is hard to find
them just from the diff. So instead print the data structures from
python too, so that they can be compared directly.

* dim-testsuite - reorder test output

The testoutput changed for some parts, so reorder them to make them
correct.
Also add a missing line in the output.