dim - fix pool list with attributes
This implementes the missing tests for #48.

When the user managed attribut list for pools was implemented it was
tested with the default field list (vlan, name), subnets and custom
attributes, but never with the standard attributes.
Needless to say, they didn't work, because they were never selected in
the query and therefore were always empty.

This commit fixes that situation and at the same time adds the missing
tests.