dim - fix allocator for ipv6
When the IPv6 split was computed the datatype changed to float. This
caused the loss of the smallest bits because of how float was
implemented.
Using integer division this problem was partially circumvented. But then
the addition of two IPv6 addresses as integers will cause the result to
be of type long. This is a problem as integer division doesn't seem to be
implemented for long. The result was always 0, but didn't cause a type
error.

This means we needed to change the formula a bit, to run the calculation
with smaller numbers.