ndcli,dim-testsuite: add support for NDCLI_USERNAME and NDCLI_SERVER
By adding environment variables NDCLI_USERNAME and NDCLI_SERVER ndcli
can not be configured to use other servers or usernames via environment
variables too.

NDCLI_COOKIEPATH was also added if the cookie path for the same user
needs to be adjusted.

The default cookiepath was changed from ~/.ndcli.cookie to
~/.ndcli.cookie.{ username }.
Cookies are already stored per hostname in the cookie store, but there
is no distinction between different usernames. Therefore it was possible
that for a different user the same cookie was used.
By changing the default, each username now gets its own cookie store and
mix ups do not happen anymore.

There is the drawback for people who test with many different users,
thay they will accumulate many different files now.

Another problem is, that this change brough to light a whole bunch of
new errors in the dim testsuite.
Some tests changed between different users, which didn't work properly
in the past. These are now all failing as sometimes users are not
created or the wrong user is used for the test.

The testsuite was adjusted to support the new flags with the name names.
Using NDCLI_USERNAME and NDCLI_SERVER for the default tests is working.