ndcli - allow override of ndclirc
For debugging it is a hindrance to pull in the default `~/.ndclirc` as
it might contain overrides for some variables.

To avoid the problem a new environment variable NDCLI_CONFIG is defined
to point ndcli somewhere else for the config. This is also the default
for running the tests and can be overriden with the same environment
variable.

This would also allow users to have multiple configs for each
layer3domain they are using and prepare them as aliases, for example

  alias ndcli_orgA=NDCLI_CONFIG=~/.ndclirc_orgA ndcli
  alias ndcli_orgB=NDCLI_CONFIG=~/.ndclirc_orgB ndcli