add API to move container to layer3domain
fixes #124

Add the API and ndcli command to move a container to a new layer3domain.
All static entries directly underneath the container will be moved to
too. All other entries (subnets, reserved IPs, ...) will be kept und
attached to the parent container.
When no parent container exists the action will be aborted as it would
break the assumption that a subnet has a parent.