dim-testsuite: add separate make targets
* add more reusable targets and add some dependencies
* add a separate target for editable pip install
* add target for pytest
* stricter requirements for deps

Adding a separate target for the editable pip install,
as that causes issues with e.g. read-only bind mounts
of the code in containers, as editable install adds
files to the *source* tree.

Adding the `pytest` suite to run the unit tests,
add default `LAYER3DOMAIN_WHITELIST` when running
with `testing` enabled, so the `pytest` tests can run.

Specifying stricter versions so `pip` can resolve the
dependencies in a timely fashion.
It complained
```This is taking longer than usual.
You might need to provide the dependency resolver
with stricter constraints to reduce runtime.
```
and would not complete within 15 minutes.