ndcli - add global dryrun flag
The dryrun flag is enabled globally but needed handling in every CLI
function.
This changes it in such a way, that the flag is introduced globally to
the request. Now all create, modify, delete and rename requests
automatically will receive the flag when it is set on the command line.

Fixes #87