dim - allow network and dns permissions on same group
Fixes #171

A user had to be in different groups to receive network and dns
permissions.
With this change a user can be in a single group and receive network and
dns permissions.

This should also work for groups in departments. As these can't be
created via DIM, this can not be tested though.