dim - allow NULL MX records (RFC7505)
To allow NULL MX records it was necessary to extend the validate_args
method with the context of other fields.

According to RFC7505 a preference of 0 is allowed when the target is
only '.' and the same in the other direction.
That means that both fields need to be checked in such cases which
requires the context of the other field.

In the future this allows us to implement similar validations for other
record types.

Fixes #128