fix #238 - do not allow A/AAAA/PTR records in zone-profiles
When zone-profiles are copied to become a real zone, the A and AAAA
records must belong to layer3domain.

It is not feasible to specify on the layer3domain each A/AAAA record
needs to go to when ndcli create zone is executed.

To avoid this problem, zone-profiles must not contain A or AAAA records.